Refactor the createJsonRpcClient & createInfuraClient tests into 1
